Set `suppress_errors=false` in Franklin build
We broke the JuliaCon website on the first day because an error was suppressed.  The error had been live for a while, but a change to surrounding code caused it to move from minor annoyance to catastrophic.  This moves us to a stricter mode where such errors are not allowed at all.